sdks/python/sdk/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "moss"
version = "1.0.0b19"
description = "Python SDK for semantic search with on-device AI capabilities"
readme = "README.md"
license-files = ["LICENSE"]
authors = [
    { name = "InferEdge Inc.", email = "contact@moss.dev" }
]
keywords = ["search", "semantic", "embeddings", "vector", "usemoss", "moss"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "typing-extensions>=4.0.0",
    "httpx>=0.25.0",
    "inferedge-moss-core==0.8.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "tox>=4.0.0",
    "black>=25.9.0",
    "isort>=7.0.0",
    "flake8>=7.3.0",
    "mypy>=1.18.2",
    "python-dotenv>=1.0.0",
    "pdoc>=15.0.4",
    "pydoc-markdown>=4.8.2",
    "griffe>=0.40.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/usemoss/moss"
Repository = "https://github.com/usemoss/moss"
Documentation = "https://docs.usemoss.dev/"
Source = "https://github.com/usemoss/moss/tree/main/sdks/python/sdk"
Changelog = "https://github.com/usemoss/moss/blob/main/sdks/python/sdk/CHANGELOG.md"

# Note: Documentation is generated via: python scripts/generate_docs.py
# [project.scripts]
# docs = "pydoc_markdown.main:cli"  # Replaced with Griffe-based script

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.package-data]
"moss" = [
    "py.typed",
    "__init__.pyi",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
